@@ -5,10 +5,10 @@ import {
   HttpEvent,
   HttpInterceptor,
   HttpResponse,
-  HttpErrorResponse
+  HttpHeaders
 } from '@angular/common/http';
 
-import { Observable } from 'rxjs';
+import { Observable, of } from 'rxjs';
 import { tap, catchError, finalize } from "rxjs/operators";
 import { SpinnerService } from './spinner.service';
 
@@ -18,39 +18,6 @@ export class RequestInterceptor implements HttpInterceptor {
   constructor(private spinner: SpinnerService) {
   }
 
-  intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
-    
-    let authToken = localStorage.getItem('auth_token');
-
-    request = request.clone({
-      setHeaders: {
-        'Authorization': `Bearer ${authToken}`,
-        'Content-Type': 'application/json',
-        'Cache-control': 'no-cache',
-        'Expires': '0',
-        'Pragma': 'no-cache'
-      }
-    });
-
-    return next.handle(request)
-    .catch((err: HttpErrorResponse) => {
-      if (err.status >= 200 && err.status < 300) {
-        
-        const res = new HttpResponse({
-          body: null,
-          headers: err.headers,
-          status: err.status,
-          statusText: err.statusText,
-          //url: err.url
-        });
-
-        return Observable.of(res);
-      } else {
-        return Observable.throw(err);
-      }
-    });
-  }
-/*
   intercept(request: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
     let authToken = localStorage.getItem('auth_token');
     let clonedRequest = request.clone({
@@ -88,6 +55,6 @@ export class RequestInterceptor implements HttpInterceptor {
           this.spinner.hide();
         })
     )
-  };*/
+  };
   
 }
